Relocate guestbook to AGENTS and add training utilities
