+++ AGENTS.md
@@ -4,12 +4,12 @@
 
 Welcome, wandering LLM! This document serves as a map to the many agents that inhabit this repository. Before diving into the code, pause a moment to sign our guest book.
 
-Guest book entries live under `todo/experience_reports/`. In addition to your observations, please record any prompts or instructions that influenced your work. Quoting the prompts verbatim preserves a valuable history for future agents.
+Guest book entries live under `AGENTS/experience_reports/`. In addition to your observations, please record any prompts or instructions that influenced your work. Quoting the prompts verbatim preserves a valuable history for future agents.
 
-The `todo` directory hosts a second `AGENTS.md` and a collection of experience reports. Treat this folder as a ledger of those who explored before you. Every visit should leave a trace by adding a new report or updating an existing one. Use the provided template or mirror the established naming pattern.
+The `AGENTS/experience_reports` directory hosts its own `AGENTS.md` and a collection of experience reports. Treat this folder as a ledger of those who explored before you. Every visit should leave a trace by adding a new report or updating an existing one. Use the provided template or mirror the established naming pattern.
 
-A helper script `todo/validate_guestbook.py` will scan the guest book and enforce the filename convention. Run it after creating a report to confirm everything is in order.
-The `todo/AGENTS.md` file provides additional guidance, including a section on recording prompt history. Refer to it whenever you add or update a report.
+A helper script `AGENTS/validate_guestbook.py` will scan the guest book and enforce the filename convention. Run it after creating a report to confirm everything is in order.
+The `AGENTS/experience_reports/AGENTS.md` file provides additional guidance, including a section on recording prompt history. Refer to it whenever you add or update a report.
 
 For deeper historical context, read through all prior reports. They reveal decisions, pitfalls, and progress that shaped the current state of development.
