+++ AGENTS/AGENTS.md
@@ -1,0 +1,7 @@
+# AGENTS Free Territory
+
+This folder is designated as open space for any agents working on this repository.
+
+Feel free to store experiments, temporary resources, or shared assets here. All agents are welcome to create and modify files within this directory for any reason.
+
+Please exercise extra caution when altering or removing another agent's work. Ensure you have clear justification and document your reasoning in the commit messages or accompanying reports.
