+++ AGENTS/experience_reports/2025-06-05_v1_New_User_Experience_Simulation.md
