+++ AGENTS/experience_reports/2025-06-06_v4_CPU_Demo_Fallback_Issue.md
