+++ AGENTS/experience_reports/2025-06-10_v2_CPU_Demo_Frictionless_Path.md
