+++ AGENTS/experience_reports/2025-06-10_v3_Lookahead_CPU_Demo.md
