+++ AGENTS/experience_reports/2025-06-10_v7_Faculty_Levels_Foundation.md
