+++ AGENTS/experience_reports/2025-06-11_v3_Free_Agent_Territory.md
@@ -1,0 +1,23 @@
+# Template User Experience Report
+
+**Date/Version:** 2025-06-11 v3
+**Title:** Free Agent Territory Folder
+
+## Overview
+Added a new `AGENTS` folder to provide open space for experimentation. Documented usage guidelines within `AGENTS/AGENTS.md`.
+
+## Prompts
+"Please make an AGENTS folder, in it an agents file, in it explain that this folder is free territory for any agents for any reason and to ask to be exceptionally careful about undoing anything anyone else has done without justification and record."
+
+## Steps Taken
+1. Created `AGENTS/AGENTS.md` describing the purpose and etiquette for the new directory.
+2. Recorded this change in a new experience report.
+
+## Observed Behaviour
+Folder and file created successfully with no errors.
+
+## Lessons Learned
+The guestbook process helps track small but important additions. Keeping a written record of prompts clarifies intent.
+
+## Next Steps
+Monitor how future agents use the folder and update documentation if new norms emerge.
