+++ AGENTS/experience_reports/archive/2025-06-06_v10_Scorer_Docstring_Refactor.md
