+++ AGENTS/experience_reports/archive/2025-06-06_v2_Torch_Bypass_Plan.md
