+++ AGENTS/experience_reports/archive/2025-06-06_v3_CPU_Demo_Mode.md
