+++ AGENTS/experience_reports/archive/2025-06-06_v6_Windows_No_PowerShell_Instructions.md
