+++ AGENTS/experience_reports/archive/2025-06-06_v7_Beam_Search_Abstraction_Study.md
