+++ AGENTS/experience_reports/archive/2025-06-06_v8_Generic_Tensor_Wrapper.md
