+++ AGENTS/experience_reports/archive/2025-06-06_v9_Reinstall_Script_Automation.md
