+++ AGENTS/experience_reports/archive/2025-06-07_v3_Missing_Transformers_Module.md
