+++ AGENTS/experience_reports/archive/2025-06-07_v4_Torch_Transformers_Flag_Issue.md
