+++ AGENTS/experience_reports/archive/2025-06-08_v1_MetaBeamManager_Agnostic_Eval.md
