+++ AGENTS/experience_reports/archive/2025-06-08_v2_MetaBeamManager_Dissolution.md
