+++ AGENTS/experience_reports/archive/2025-06-08_v3_Scorer_Bins_Integration.md
